refactor instant quote page to server wrapper
